requirements.txt
asgiref==3.5.2
Django==4.1.3
django-js-asset==2.0.0
django-mptt==0.14.0
django-multiselectfield==0.1.12
Pillow==9.3.0
sqlparse==0.4.3
tzdata==2022.7
